Add transactions for speed

+++ bin/update_holdings.ml
@@ -20,20 +20,21 @@ let rec handle_sell db sale_id n_quantity_to_sell =
       let isin = (List.hd_exn !rows).(0) |> Option.value ~default:"" in
       let buyrows = ref [] in
       (* Find current holding lots of this stock *)
-      if Db_wrapper.run_query_callback db ~cb:(fun x -> buyrows := x::!buyrows) ("select lot_id, n_stocks from current_holdings join raw_transaction_information on raw_transaction_information.id = lot_id where isin = '" ^ isin ^ "' order by id desc;")
+      if Db_wrapper.run_query_callback db ~cb:(fun x -> buyrows := x::!buyrows) ("select lot_id, n_stocks from current_holdings join raw_transaction_information on raw_transaction_information.id = lot_id where isin = '" ^ isin ^ "' order by id desc;") (* && Db_wrapper.run_query db "begin transaction" *)
       then
         match List.hd_exn !buyrows with
         | [|Some buy_id;Some quantity|] ->
-           let quantity = Float.of_string quantity in
-           if Float.(quantity > n_quantity_to_sell) then
-             if Db_wrapper.run_query db (Printf.sprintf "insert into sale_data (sale_id, buy_id, n_stocks) values (%s, %s, %f)" sale_id buy_id n_quantity_to_sell) && (Db_wrapper.run_query db (Printf.sprintf "update current_holdings set n_stocks = %f where lot_id = %s;" (quantity -. n_quantity_to_sell) buy_id)) then
-               ()
-             else
-               print_endline ("Error importing sale transaction for " ^ sale_id ^ " (" ^ isin ^ ")!")
-           else
-             if Db_wrapper.run_query db (Printf.sprintf "insert into sale_data (sale_id, buy_id, n_stocks) values (%s, %s, %f)" sale_id buy_id quantity) && Db_wrapper.run_query db ("delete from current_holdings where lot_id = " ^ buy_id ^ ";") then
-               handle_sell db sale_id (n_quantity_to_sell -. quantity)
+          let quantity = Float.of_string quantity in
+          if Float.(quantity > n_quantity_to_sell) then
+            if Db_wrapper.run_query db (Printf.sprintf "insert into sale_data (sale_id, buy_id, n_stocks) values (%s, %s, %f)" sale_id buy_id n_quantity_to_sell) && (Db_wrapper.run_query db (Printf.sprintf "update current_holdings set n_stocks = %f where lot_id = %s;" (quantity -. n_quantity_to_sell) buy_id)) then
+              ()
+            else
+              print_endline ("Error importing sale transaction for " ^ sale_id ^ " (" ^ isin ^ ")!")
+          else
+          if Db_wrapper.run_query db (Printf.sprintf "insert into sale_data (sale_id, buy_id, n_stocks) values (%s, %s, %f)" sale_id buy_id quantity) && Db_wrapper.run_query db ("delete from current_holdings where lot_id = " ^ buy_id ^ ";") then
+            handle_sell db sale_id (n_quantity_to_sell -. quantity)
         | _ -> print_endline ("Failure in sale transaction " ^ sale_id ^ "!")
+          (* if Db_wrapper.run_query db "end transaction" then print_endline "Insert complete!"; *)
       else
         print_endline ("Error handling sale with id " ^ sale_id ^ "!")
 
@@ -46,15 +47,19 @@ let handle_buy db id quantity =
 let handle_buy_and_sell db entry =
   match entry with
   | [|Some id;Some quantity;Some buy_or_sell|] ->
-     if (String.equal buy_or_sell "B") then handle_buy db id quantity else handle_sell db id (Float.of_string quantity)
+    if (String.equal buy_or_sell "B") then handle_buy db id quantity else handle_sell db id (Float.of_string quantity)
   | _ -> print_endline "ERROR GENERATING HOLDINGS!"
 
 let get_pending_holdings db =
-  let rows = ref [] in
-  if Db_wrapper.run_query_callback db ~cb:(fun x -> rows := x::!rows) "select id, quantity, buy_or_sell from raw_transaction_information where id not in (select distinct lot_id from current_holdings union select sale_id from sale_data) order by order_date asc, trade_num asc, id asc;" then
-    List.iter ~f:(handle_buy_and_sell db) (List.rev !rows)
+  if Db_wrapper.run_query db "begin transaction" then
+    let rows = ref [] in
+    if Db_wrapper.run_query_callback db ~cb:(fun x -> rows := x::!rows) "select id, quantity, buy_or_sell from raw_transaction_information where id not in (select distinct lot_id from current_holdings union select sale_id from sale_data) order by order_date asc, trade_num asc, id asc;" then
+      (List.iter ~f:(handle_buy_and_sell db) (List.rev !rows);
+       if Db_wrapper.run_query db "end transaction" then () else print_endline "ERROR ending transaction!")
+    else
+      print_endline "ERROR"
   else
-    print_endline "ERROR"
+    print_endline "ERROR initiating transaction!"
 
 
 let process_with_db db =
